move banSvair to ban directory
